[NOID] Cherry pick ApocConfigCommandExpansionTest fixes
