+++ core/src/test/resources/apoc-config-command-expansion/apoc.conf
@@ -1,0 +1,4 @@
+# used in ApocConfigTest command expansion tests
+# in real deployments apoc.conf should be stored in neo4j's conf folder
+foo=bar
+command.expansion=$(echo "expanded value")
